requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
dj-database-url
psycopg2-binary
django-cors-headers
gunicorn
